requirements.txt
jsondiff==0.2.0
scandir==1.3
click==6.6
pyyaml==3.12
